+++ lib/app/connected_playback/ConnectedPlaybackControllerSession.dart
@@ -127,6 +127,9 @@ class ConnectedPlaybackControllerSession {
 
   Future<Result<void>> requestSnapshot() => _send(controller.requestSnapshot());
 
+  Future<Result<void>> takeControl(String controlSessionId) =>
+      _send(controller.takeControl(controlSessionId));
+
   Future<Result<void>> joinSyncGroup(String groupId) =>
       _send(controller.joinSyncGroup(groupId));